+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 
+## Unreleased
+
+**Features**:
+
+- Added an Object type for Portable PDB files. ([#696](https://github.com/getsentry/symbolic/pull/696))
+
 ## 9.2.1
 
 **Fixes**:
+
 - Fixed a bug in Unreal Engine log parsing by updating the `anylog` dependency. ([#695](https://github.com/getsentry/symbolic/pull/695))
 
 ## 9.2.0